add initialization code; change endpoint and bundle name to avoid conflicts
